+++ src/main/java/com/moratan251/psitweaks/common/spells/PsitweaksSpells.java
@@ -74,8 +74,17 @@ public class PsitweaksSpells {
             PsiAPI.registerSpellPieceAndTexture(ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_aqua_cutter"), PieceTrickAquaCutter.class);
             PsiAPI.addPieceToGroup(PieceTrickAquaCutter.class, ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_aqua_cutter"), false);
 
+            PsiAPI.registerSpellPieceAndTexture(ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_blaze_ball"), PieceTrickBlazeBall.class);
+            PsiAPI.addPieceToGroup(PieceTrickBlazeBall.class, ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_blaze_ball"), false);
+
+            PsiAPI.registerSpellPieceAndTexture(ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_active_air_mine"), PieceTrickActiveAirMine.class);
+            PsiAPI.addPieceToGroup(PieceTrickActiveAirMine.class, ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_active_air_mine"), false);
+
             PsiAPI.registerSpellPieceAndTexture(ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_radiation_injection"), PieceTrickRadiationInjection.class);
             PsiAPI.addPieceToGroup(PieceTrickRadiationInjection.class, ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_radiation_injection"), false);
+
+            PsiAPI.registerSpellPieceAndTexture(ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_guillotine"), PieceTrickGuillotine.class);
+            PsiAPI.addPieceToGroup(PieceTrickGuillotine.class, ResourceLocation.fromNamespaceAndPath("psitweaks", "trick_guillotine"), false);
 
 
         });