+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 ## Changelog
 
 
-#### v0.7.3 
+#### v0.8.0 (planned)
+
+- Program Research Station Added
+    * Unlock new spell pieces by spending items, power, and time
+    * Powerful spells from existing mods also require unlocking
+    * Whether research is required can be changed in the config
+
+
+- Added Trick: Radiation Injection
+    * Exposes the target to radiation
+
+
+- Added Trick: Blaze Ball 
+    * Launches a fireball forward to attack.
+
 
 - Added tags to items